eng/requirements.txt
fastapi
uvicorn
python-dotenv
# sqlalchemy
# aiosqlite
# asyncpg
# pyjwt
# bcrypt
# python-multipart
# jinja2
# boto3
# boto3-stubs
# mypy-boto3-s3
# twilio
# greenlet
faststream[rabbit]
